Proposed some additional text to address discussion about payment structure
There was some discussion about a lower fee for e.g. students at the IETF-109 meeting. This should not be covered in this document as it is a separate point and in this document we want to let the LLC set the fee structure. However, this is proposing some additional sentences to make that explicit. Not sure if this additional text is needed but it also might not hurt. There was another comment that we should be positive about the expectation that people will usually pay. However, I think this is covered here and was covered before by the first sentence.